Refactor out FormNumberInput and FormSelectInput
